+++ src/content/panorama/scripts/custom_game/game_setup/game_setup_options.js
@@ -341,8 +341,10 @@ var basicOptions = {
                             'lodMutatorGottaGoFastOff': 0,
                         },
                         states: {
-                            'lodMutatorGottaGoFastOn': 1,
-                            'lodMutatorGottaGoREALLYFast': 2
+                            'lodMutatorGottaGoQuickOn': 1,
+                            'lodMutatorGottaGoFastOn': 2,
+                            'lodMutatorGottaGoREALLYFast': 3,
+                            'lodMutatorGottaGoSlow': 4
                         }
                     },
                     {
@@ -1264,12 +1266,20 @@ var advancedOptions = {
                         value: 0
                     },
                     {
-                        text: 'lodOptionGottaGoFastOn',
+                        text: 'lodOptionGottaGoQuickOn',
                         value: 1
                     },
                     {
-                        text: 'lodOptionGottaGoREALLYFast',
+                        text: 'lodOptionGottaGoFastOn',
                         value: 2
+                    },
+                    {
+                        text: 'lodOptionGottaGoREALLYFast',
+                        value: 3
+                    },
+                    {
+                        text: 'lodOptionGottaGoSlow',
+                        value: 4
                     },
                 ]
             },