+++ src/game/scripts/vscripts/abilities/global_mutators/global_mutator.lua
@@ -7,6 +7,10 @@ local OptionManager = require('optionmanager')
 -- Intrinsic modifier on the unit
 LinkLuaModifier( "modifier_global_mutator", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
 -- Gotta Go Fast modifiers
+LinkLuaModifier( "modifier_gottagoslow_aura", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
+LinkLuaModifier( "modifier_gottagoslow_effect", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
+LinkLuaModifier( "modifier_gottagoquick_aura", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
+LinkLuaModifier( "modifier_gottagoquick_effect", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
 LinkLuaModifier( "modifier_gottagofast_aura", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
 LinkLuaModifier( "modifier_gottagofast_effect", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
 LinkLuaModifier( "modifier_gottagoreallyfast_aura", "abilities/global_mutators/global_mutator.lua" ,LUA_MODIFIER_MOTION_NONE )
@@ -58,9 +62,13 @@ if IsServer() then
 	function modifier_global_mutator:OnCreated()
 		-- Gotta Go Fast
 		if OptionManager:GetOption("gottaGoFast") == 1 then
-			local thinker = CreateModifierThinker(self:GetParent(),self:GetAbility(),"modifier_gottagofast_aura",{},Vector(0,0,0),20,false)
+			local thinker = CreateModifierThinker(self:GetParent(),self:GetAbility(),"modifier_gottagoquick_aura",{},Vector(0,0,0),20,false)
 		elseif OptionManager:GetOption("gottaGoFast") == 2 then
+			local thinker = CreateModifierThinker(self:GetParent(),self:GetAbility(),"modifier_gottagofast_aura",{},Vector(0,0,0),20,false)
+		elseif OptionManager:GetOption("gottaGoFast") == 3 then
 			local thinker = CreateModifierThinker(self:GetParent(),self:GetAbility(),"modifier_gottagoreallyfast_aura",{},Vector(0,0,0),20,false)
+		elseif OptionManager:GetOption("gottaGoFast") == 4 then
+			local thinker = CreateModifierThinker(self:GetParent(),self:GetAbility(),"modifier_gottagoslow_aura",{},Vector(0,0,0),20,false)
 		end
 		-- Memes Redux
 		if OptionManager:GetOption("memesRedux") == 1 then
@@ -68,6 +76,116 @@ if IsServer() then
 		end
 	end
 end
+--------------------------------------------------------------------------------------------------------
+--		Modifier: modifier_gottagoslow_aura				
+--------------------------------------------------------------------------------------------------------
+if modifier_gottagoslow_aura ~= "" then modifier_gottagoslow_aura = class({}) end
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:IsPassive()
+	return true
+end
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:IsHidden()
+	return true
+end
+
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:RemoveOnDeath()
+	return false
+end
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:IsAura() return true end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:GetModifierAura()	return "modifier_gottagoslow_effect" end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:GetAuraSearchTeam() return DOTA_UNIT_TARGET_TEAM_ENEMY end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:GetAuraSearchType() return DOTA_UNIT_TARGET_HERO end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:GetAuraSearchFlags() return DOTA_UNIT_TARGET_FLAG_MAGIC_IMMUNE_ENEMIES end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_aura:GetAuraRadius() return 80000 end
+----------------------------------------------------------------------------------------------------------
+--------------------------------------------------------------------------------------------------------
+--		Modifier: modifier_gottagoslow_effect				
+--------------------------------------------------------------------------------------------------------
+if modifier_gottagoslow_effect ~= "" then modifier_gottagoslow_effect = class({}) end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_effect:IsDebuff()
+	return false
+end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_effect:GetTexture()
+	return "custom/mutator_gottagoslow"
+end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_effect:DeclareFunctions()
+	local funcs = {
+		MODIFIER_PROPERTY_MOVESPEED_BONUS_PERCENTAGE,
+	}
+	return funcs
+end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoslow_effect:GetModifierMoveSpeedBonus_Percentage()
+	return -25
+end
+
+
+--------------------------------------------------------------------------------------------------------
+--		Modifier: modifier_gottagoquick_aura				
+--------------------------------------------------------------------------------------------------------
+if modifier_gottagoquick_aura ~= "" then modifier_gottagoquick_aura = class({}) end
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:IsPassive()
+	return true
+end
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:IsHidden()
+	return true
+end
+
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:RemoveOnDeath()
+	return false
+end
+--------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:IsAura() return true end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:GetModifierAura()	return "modifier_gottagoquick_effect" end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:GetAuraSearchTeam() return DOTA_UNIT_TARGET_TEAM_ENEMY end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:GetAuraSearchType() return DOTA_UNIT_TARGET_HERO end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:GetAuraSearchFlags() return DOTA_UNIT_TARGET_FLAG_MAGIC_IMMUNE_ENEMIES end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_aura:GetAuraRadius() return 80000 end
+----------------------------------------------------------------------------------------------------------
+--------------------------------------------------------------------------------------------------------
+--		Modifier: modifier_gottagoquick_effect				
+--------------------------------------------------------------------------------------------------------
+if modifier_gottagoquick_effect ~= "" then modifier_gottagoquick_effect = class({}) end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_effect:IsDebuff()
+	return false
+end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_effect:GetTexture()
+	return "custom/mutator_gottagoquick"
+end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_effect:DeclareFunctions()
+	local funcs = {
+		MODIFIER_PROPERTY_MOVESPEED_BONUS_PERCENTAGE,
+	}
+	return funcs
+end
+----------------------------------------------------------------------------------------------------------
+function modifier_gottagoquick_effect:GetModifierMoveSpeedBonus_Percentage()
+	return -15
+end
+
+
 --------------------------------------------------------------------------------------------------------
 --		Modifier: modifier_gottagofast_aura				
 --------------------------------------------------------------------------------------------------------
@@ -108,7 +226,7 @@ function modifier_gottagofast_effect:IsDebuff()
 end
 ----------------------------------------------------------------------------------------------------------
 function modifier_gottagofast_effect:GetTexture()
-	return "mutator_gottagofast"
+	return "custom/mutator_gottagofast"
 end
 ----------------------------------------------------------------------------------------------------------
 function modifier_gottagofast_effect:DeclareFunctions()
@@ -171,7 +289,7 @@ function modifier_gottagoreallyfast_effect:IsDebuff()
 end
 ----------------------------------------------------------------------------------------------------------
 function modifier_gottagoreallyfast_effect:GetTexture()
-	return "mutator_gottagoreallyfast"
+	return "custom/mutator_gottagoreallyfast"
 end
 ----------------------------------------------------------------------------------------------------------
 function modifier_gottagoreallyfast_effect:DeclareFunctions()